app/pyproject.toml
[tool.poetry]
name = "itps-research"
version = "0.5.0"
description = "The ITPS Research Database"
authors = ["Jason Heppler <jason.heppler@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
gunicorn = "^23.0.0"
psycopg2 = "^2.9.10"
django-environ = "^0.11.2"
python-dotenv = "^1.0.1"
geopy = "^2.4.1"
django-tables2 = "^2.7.0"
django-filter = "^24.3"
django-bootstrap5 = "^24.3"
django-tailwind = "^3.8.0"
pillow = "^11.1.0"
django-import-export = "^4.3.7"
xlsxwriter = "^3.2.2"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
djhtml = "^3.0.7"
pre-commit = "^4.0.1"
django-debug-toolbar = "^4.4.6"
pytest-django = "^4.9.0"
pylint-django = "^2.6.1"


[tool.poetry.group.development.dependencies]
django-tailwind = {extras = ["reload"], version = "^3.8.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
